Allow users to set the openproject host name without using the ingress (#106)
* Allow users to set the openproject host name without using the ingress

Right now, you must have the ingress enabled in order to set the hostname of openproject. This change to the secret core allows the host to specified host value even when ingress is disabled. This is beneficial to users who are using the Gateway API HTTPRoutes instead of traditional Ingresses

* Update values.yaml to clarify openproject host is defaulted to ingress value when ingress is enabled

* Create sweet-laws-draw.md

---------

Co-authored-by: Oliver Günther <mail@oliverguenther.de>